+++ src/extension.ts
@@ -1,26 +1,26 @@
 import * as vscode from 'vscode';
 import { FileComparator } from './fileComparator';
 import { DiffViewProvider } from './diffViewProvider';
-import { MeldenUriHandler } from './uriHandler';
+import { BygoneUriHandler } from './uriHandler';
 
 export function activate(context: vscode.ExtensionContext) {
     const fileComparator = new FileComparator();
     const diffViewProvider = new DiffViewProvider(context.extensionUri);
-    const uriHandler = new MeldenUriHandler(fileComparator);
-    const standaloneDownloadUrl = vscode.Uri.parse('https://github.com/davidmashburn/melden/releases');
+    const uriHandler = new BygoneUriHandler(fileComparator);
+    const standaloneDownloadUrl = vscode.Uri.parse('https://github.com/davidmashburn/bygone/releases');
 
     fileComparator.setDiffViewProvider(diffViewProvider);
 
     context.subscriptions.push(
         vscode.window.registerWebviewViewProvider(DiffViewProvider.viewType, diffViewProvider),
         vscode.window.registerUriHandler(uriHandler),
-        registerCommand('melden.compareFiles', () => fileComparator.selectAndCompareFiles()),
-        registerCommand('melden.compareWithSelected', (resource: vscode.Uri) => fileComparator.compareWithSelected(resource)),
-        registerCommand('melden.threeWayMerge', () => fileComparator.threeWayMerge()),
-        registerCommand('melden.compareTestFiles', () => fileComparator.compareTestFiles()),
-        registerCommand('melden.compareFileHistory', (resource?: vscode.Uri) => fileComparator.compareFileHistory(resource)),
-        registerCommand('melden.compareActiveFileHistory', () => fileComparator.compareFileHistory()),
-        registerCommand('melden.openStandaloneDownloads', () => vscode.env.openExternal(standaloneDownloadUrl))
+        registerCommand('bygone.compareFiles', () => fileComparator.selectAndCompareFiles()),
+        registerCommand('bygone.compareWithSelected', (resource: vscode.Uri) => fileComparator.compareWithSelected(resource)),
+        registerCommand('bygone.threeWayMerge', () => fileComparator.threeWayMerge()),
+        registerCommand('bygone.compareTestFiles', () => fileComparator.compareTestFiles()),
+        registerCommand('bygone.compareFileHistory', (resource?: vscode.Uri) => fileComparator.compareFileHistory(resource)),
+        registerCommand('bygone.compareActiveFileHistory', () => fileComparator.compareFileHistory()),
+        registerCommand('bygone.openStandaloneDownloads', () => vscode.env.openExternal(standaloneDownloadUrl))
     );
 }