+++ standalone/preload.js
@@ -4,16 +4,16 @@ const { contextBridge, ipcRenderer } = require('electron');
 
 const editorWorkerUrl = pathToFileURL(path.join(__dirname, '..', 'media', 'editor.worker.js')).toString();
 
-ipcRenderer.on('melden:host-message', (_event, payload) => {
+ipcRenderer.on('bygone:host-message', (_event, payload) => {
     window.postMessage({
-        __meldenHostMessage: payload
+        __bygoneHostMessage: payload
     }, '*');
 });
 
-contextBridge.exposeInMainWorld('__MELDEN_HOST__', {
+contextBridge.exposeInMainWorld('__BYGONE_HOST__', {
     environment: 'standalone',
     editorWorkerUrl,
     postMessage(message) {
-        ipcRenderer.send('melden:renderer-message', message);
+        ipcRenderer.send('bygone:renderer-message', message);
     }
 });